Renaming and adding interfaces
Renaming interface Six to Eight according to the kyu of related kata. Adding interface Seven for kata of 7 kui
