+++ src/com/ss/ita/kata/Seven.java
@@ -1,0 +1,38 @@
+package com.ss.ita.kata;
+
+/**
+ * For kata of 7 kyu
+ */
+public interface Seven {
+    /**
+     * Looking for a benefactor
+     * @param arr
+     * @param navg
+     * @return
+     */
+    long newAvg(double[] arr, double navg);
+
+    /**
+     * Sum of the first nth term of Series
+     * @param n
+     * @return
+     */
+    String seriesSum(int n);
+
+    /**
+     * Where is Vasia?
+     * @param p
+     * @param bef
+     * @param aft
+     * @return
+     */
+    int whereIsHe(int p, int bef, int aft);
+
+    /**
+     * Build a pile of Cubes
+     * @param m
+     * @return
+     */
+    long findNb(long m);
+
+}
