+++ src/com/ss/ita/kata/implementation/solomiazh/EightImpl.java
@@ -1,23 +1,35 @@
-package com.ss.ita.kata.implementation.rsavu98;
+package com.ss.ita.kata.implementation.solomiazh;
 
-import com.ss.ita.kata.Six;
+import com.ss.ita.kata.Eight;
 
-public class SixImpl implements Six {
-  
+import java.text.DecimalFormat;
+
+public class EightImpl implements Eight {
+    // const
+    private static final float LITER = 4.54609188f;
+    private static final float KILOMETER = 1.609344f;
+
     @Override
-    public int liters (double time) {
-        return (int)(time / 2);
+    public int liters(double time) {
+        return (int) (time / 2);
+
     }
 
     @Override
     public double getVolumeOfCuboid(double length, double width, double height) {
-        return length * width * height;
+
+        return 0;
     }
 
     @Override
     public float mpgToKPM(float mpg) {
-        return (float)Math.round(mpg * 1.609344 / 4.54609188 * 100) / 100;
+
+        double res = (mpg * KILOMETER / LITER * 100) / 100;
+        DecimalFormat myFormatter = new DecimalFormat("###.##");
+        String output = myFormatter.format(res);
+        return Float.parseFloat(output);
     }
+
 
     @Override
     public int[] squareOrSquareRoot(int[] array) {
@@ -43,11 +55,9 @@ public class SixImpl implements Six {
     public double twoDecimalPlaces(double number) {
         return 0;
     }
-  
 
     @Override
     public int[] divisibleBy(int[] numbers, int divider) {
         return new int[0];
     }
 }
-
